+++ ydb/mvp/meta/support_links/ya.make
@@ -2,6 +2,7 @@ LIBRARY()
 
 SRCS(
     events.h
+    resolve_types.h
     source.h
     source.cpp
     support_links_resolver.h
